+++ spec/features/stories_spec.rb
@@ -48,6 +48,28 @@ describe "Stories" do
 
   end
 
+  describe "story history" do
+
+    before do
+      create(:story, title: 'Test Story', project: project, requested_by: user)
+      visit project_path(project)
+      wait_spinner
+      wait_page_load
+
+      find('.story-title').click
+      find('.toggle-history').click
+    end
+
+    it "turns history visible", js: true do
+      expect(page).to have_css('#history')
+    end
+
+    it "updates history column title", js: true do
+      title = find('.history_column > .column_header > .toggle-title')
+      expect(title).to have_text("History Of 'Test Story'")
+    end
+  end
+
   describe "story links" do
 
     let!(:story) { create(:story, title: "Story", project: project, requested_by: user)}