+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - tag group form
 - Added First time tour.
 
+### Changed
+- Story id, location and history input-form-group into a react component 
+
 ### Fixed
 - Remove the API module used in `using` params into Project Entity
 - Clean up DatabaseCleaner config